@akihikodaki 2017-06-22
301c2857b4bf4fff32a8302b77151d52be0e5c99 runs migration in the concurrent way.

> In this case you won't want to remove the original index until after the concurrent one is created, so maybe it should wait for a followup release.

If I recall correctly, queries in a migration runs in a transaction, so it would wait for the previous query finishes.