@nightpool 2017-06-22
@akihikodaki right, but concurrently created queries *can't* run in a transaction. That's what the disable_ddl_transaction does.