@akihikodaki 2017-06-22
@nightpool Ah, I forgot that I did `disable_ddl_transaction!`. But still queries in a migration would be queued serially in a connection and `CREATE INDEX CONCURRENTLY` would block in the connection, I think.
`CREATE INDEX CONCURRENTLY` does not lock the database, but does not it block the connection?