@nightpool 2017-06-22
@akihikodaki uh, I'm not actually 100% sure. I thought that it did, but I might be wrong.