@nightpool 2017-06-22
yeah, nevermind that was just me being stupid. Of course it waits for the index to finish building before returning.