@akihikodaki 2017-06-22
That is fine. Could you approve or tell me other concerns if any?