@nightpool 2017-06-22
I dismissed my review, and I think the PR is good, but I'm a very new contributor and this is pretty sensitive code (could break upgrades entirely for people) so I'm not comfortable approving it without someone who has more experience with the upgrade process and handling admin concerns looking at it.